requirements.txt
mypy
flake8
black
isort
crc32c
python-dotenv
requests
pydantic
pyproj
